requirements.txt
python-dotenv
semantic-router==0.0.20
langchain==0.0.352
openai>=1.6.1
langchain-openai